Special abilities of magicians
Added special abilities of magicians

+++ src/classes/characters/AbstractCharacter.java
@@ -6,19 +6,19 @@ import java.util.Random;
  * Created by Vlad Yakovenko on 08.07.2017.
  */
 public abstract class AbstractCharacter {
-    protected int hitPoints = 100;
+    protected double hitPoints = 100;
     protected Random rand = new Random();
 
-    public void gotDamage(int damage) {
+    public void gotDamage(double damage) {
         hitPoints -= damage;
         showDamage(damage);
     }
 
-    protected void showDamage(int damage) {
+    protected void showDamage(double damage) {
         System.out.println(" inflicted " + damage + "hp damage to " + this.getClassName());
     }
 
-    public int getHitPoints() {
+    public double getHitPoints() {
         return hitPoints;
     }
 
@@ -28,5 +28,5 @@ public abstract class AbstractCharacter {
         return name;
     }
 
-    public abstract int goDamage();
+    public abstract double goDamage();
 }
