+++ src/classes/characters/elves/ElfArcher.java
@@ -18,7 +18,7 @@ public class ElfArcher extends AbstractCharacter implements Archer{
     }
 
     @Override
-    public int goDamage() {
+    public double goDamage() {
         System.out.print(getClassName());
         if((rand.nextInt(2)+1)==1)
             return attack();
