+++ src/classes/characters/elves/ElfMagician.java
@@ -18,7 +18,7 @@ public class ElfMagician extends AbstractCharacter implements Magician {
     }
 
     @Override
-    public int goDamage() {
+    public double goDamage() {
         System.out.print(getClassName());
         if((rand.nextInt(2)+1)==1)
             return improveCharacter();
