+++ src/classes/characters/people/Crossbowman.java
@@ -17,7 +17,7 @@ public class Crossbowman extends AbstractCharacter implements Archer{
     }
 
     @Override
-    public int goDamage() {
+    public double goDamage() {
         System.out.print(getClassName());
         if((rand.nextInt(2)+1)==1)
             return attack();
