+++ src/classes/characters/people/ManWarrior.java
@@ -12,7 +12,7 @@ public class ManWarrior extends AbstractCharacter implements Warrior {
     }
 
     @Override
-    public int goDamage() {
+    public double goDamage() {
         System.out.print(getClassName());
         return attackWithWeapons();
     }
